Code cleanup
Removed ~80 unused assets, fixed styling issues with quotes, variable
declarations, spacing, semicolons, etc. according to linter, and cleaned
up comments and general code style.

The assets in the public folder for this commit came from my fork of
the current map, and as such the asset id numbers will correspond with
the ids from that project:

https://playcanvas.com/project/496661/overview/third-current-ballhalla-james